tornado/requirements.txt
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
Django==1.8.4
django-rest-framework-mongoengine==2.0.2
certifi==2015.04.28
flup==1.0.2
mongoengine==0.9.0
pymongo==2.8.1
pytz==2015.4
requests==2.7.0
tornado==4.2
wsgiref==0.1.2
djangorestframework==3.2.3
